Add tests in separate file testing.js
