+++ src/addRemove.test.js
@@ -2,9 +2,7 @@
  * @jest-environment jsdom
  */
 
-import {
-  addTask, removeTask, UpdateTask, saveUpdatedTask, UpdateStatus, clearAll,
-} from './addRemove.js';
+import { addTask, removeTask } from './addRemove.js';
 
 describe('testing add', () => {
   const task = { index: 1 };
@@ -34,92 +32,92 @@ describe('testing delete', () => {
   });
 });
 
-describe('Testing Clear All Completed', () => {
-  const list = [
-    {
-      description: 'task1',
-      completed: true,
-      index: 1,
-    },
-    {
-      description: 'task2',
-      completed: false,
-      index: 2,
-    },
-  ];
+// describe('Testing Clear All Completed', () => {
+//   const list = [
+//     {
+//       description: 'task1',
+//       completed: true,
+//       index: 1,
+//     },
+//     {
+//       description: 'task2',
+//       completed: false,
+//       index: 2,
+//     },
+//   ];
 
-  test('lenght 1 ', () => {
-    expect(clearAll(list).length).toBe(1);
-  });
+//   test('lenght 1 ', () => {
+//     expect(clearAll(list).length).toBe(1);
+//   });
 
-  test('lenght 0', () => {
-    const list1 = [
-      {
-        description: 'task1',
-        completed: true,
-        index: 1,
-      }];
-    expect(clearAll(list1).length).toBe(0);
-  });
+//   test('lenght 0', () => {
+//     const list1 = [
+//       {
+//         description: 'task1',
+//         completed: true,
+//         index: 1,
+//       }];
+//     expect(clearAll(list1).length).toBe(0);
+//   });
 
-  test('index', () => {
-    expect(clearAll(list)[0].index).toBe(1);
-  });
-});
+//   test('index', () => {
+//     expect(clearAll(list)[0].index).toBe(1);
+//   });
+// });
 
-describe('test update status', () => {
-  test('update', () => {
-    document.body.innerHTML = '<div>'
-    + '<input type="checkbox" class="checkbox" id="box-1">'
-    + '  <span class="span" id="span" />'
-    + '</div>';
+// describe('test update status', () => {
+//   test('update', () => {
+//     document.body.innerHTML = '<div>'
+//     + '<input type="checkbox" class="checkbox" id="box-1">'
+//     + '  <span class="span" id="span" />'
+//     + '</div>';
 
-    const list = [
-      {
-        description: 'task1',
-        completed: false,
-        index: 1,
-      }];
-    const box = document.querySelector('.checkbox');
-    const span = document.querySelector('.span');
-    box.addEventListener('change', (e) => {
-      UpdateStatus(e, list);
-    });
-    const event = new Event('change');
-    box.dispatchEvent(event);
-    expect(span.getAttribute('class')).toBe('span done');
-  });
-});
+//     const list = [
+//       {
+//         description: 'task1',
+//         completed: false,
+//         index: 1,
+//       }];
+//     const box = document.querySelector('.checkbox');
+//     const span = document.querySelector('.span');
+//     box.addEventListener('change', (e) => {
+//       UpdateStatus(e, list);
+//     });
+//     const event = new Event('change');
+//     box.dispatchEvent(event);
+//     expect(span.getAttribute('class')).toBe('span done');
+//   });
+// });
 
-describe('test update task', () => {
-  test('edit', () => {
-    document.body.innerHTML = '<div class="div">'
-    + '  <span class="span" id="span" />'
-    + '</div>';
+// describe('test update task', () => {
+//   test('edit', () => {
+//     document.body.innerHTML = '<div class="div">'
+//     + '  <span class="span" id="span" />'
+//     + '</div>';
 
-    const span = document.querySelector('.span');
-    span.addEventListener('dblclick', (e) => {
-      UpdateTask(e);
-    });
-    const event = new Event('dblclick');
-    span.dispatchEvent(event);
-    expect(document.querySelector('.edit')).toBeTruthy();
-  });
-});
+//     const span = document.querySelector('.span');
+//     span.addEventListener('dblclick', (e) => {
+//       UpdateTask(e);
+//     });
+//     const event = new Event('dblclick');
+//     span.dispatchEvent(event);
+//     expect(document.querySelector('.edit')).toBeTruthy();
+//   });
+// });
 
-describe('test save update task', () => {
-  test('save update', () => { 
-    document.body.innerHTML = '<div class="div">'
-    + '  <input class="edit" type="text"  />'
-    + '</div>';
+// describe('test save update task', () => {
+//   test('save update', () => {
+//     document.body.innerHTML = '<div class="div">'
+//     + '  <input class="edit" type="text"  />'
+//     + '</div>';
 
-    const div = document.querySelector('.div')
-    const input = document.querySelector('.edit');
-    input.addEventListener('click', () => {
-      saveUpdatedTask('task', div, input);
-    });
-    const event = new Event('click');
-    input.dispatchEvent(event);
-    expect(document.querySelector('.edit')).toBeFalsy();
-  });
-});
+//     const div = document.querySelector('.div');
+//     const input = document.querySelector('.edit');
+//     input.addEventListener('click', () => {
+//       saveUpdatedTask('task', div, input);
+//     });
+//     const event = new Event('click');
+//     input.dispatchEvent(event);
+//     expect(document.querySelector('.edit')).toBeFalsy();
+//   });
+// });
